backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
matplotlib
python-dotenv
mne
pandas
asrpy